refactor(bare-strings): rename to no-bare-strings
