+++ docs/rule/no-bare-strings.md
@@ -1,4 +1,4 @@
-## bare-strings
+## no-bare-strings
 
 In order to be able to internationalize your application, you will need to avoid using plain strings in your templates. Instead, you would need to use a template helper specializing in translation ([ember-i18n](https://github.com/jamesarosen/ember-i18n) and [ember-intl](https://github.com/yahoo/ember-intl) are great projects to use for this).
