+++ lib/rules/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 module.exports = {
-  'bare-strings': require('./lint-bare-strings'),
+  'no-bare-strings': require('./lint-no-bare-strings'),
   'block-indentation': require('./lint-block-indentation'),
   'no-debugger': require('./lint-no-debugger'),
   'no-log': require('./lint-no-log'),
