+++ lib/rules/lint-no-bare-strings.js
