+++ test/fixtures/with-errors/.template-lintrc.js
@@ -1,5 +1,5 @@
 module.exports = {
   rules: {
-    'bare-strings': true
+    'no-bare-strings': true
   }
 };
