+++ test/fixtures/with-plugin-with-configurations/plugins/plugin1.js
@@ -15,7 +15,7 @@ module.exports = {
 
       rules: {
         'inline-component': true,
-        'bare-strings': true
+        'no-bare-strings': true
       }
 
     }
