+++ test/fixtures/with-plugins/.template-lintrc.js
@@ -4,6 +4,6 @@ module.exports = {
   ],
   rules: {
     'inline-component': true,
-    'bare-strings': true
+    'no-bare-strings': true
   }
 };
