+++ test/unit/rules/lint-block-indentation-test.js
@@ -181,7 +181,7 @@ generateRuleTests({
         '\t<p>Hi!</p>\n' +
         '</div>'
     },
-    '{{! template-lint-disable bare-strings }}',
+    '{{! template-lint-disable no-bare-strings }}',
     '{{! template-lint-disable }}',
     {
       template: [
